Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
tensorflow
numpy==1.26.4  # Let's lock the version that was successfully installed
Pillow
python-multipart
opencv-python
mediapipe